First rough implementation of the swapmesh tool.
Missing Features:
  * include OpenMP (commands are there but library is not included)
  * 2D support
  * FV

The no longer necessary routines to get the length of the ini file for
the user block has been removed with the global variable userblock_len.

Renamed ParametricCoordinates modules in recordpoints and swapmesh to
avoid confusion.
Moved calculation of determinant and inverse of 2x2 and 3x3 matrices to
mathtools.

For the future, the search for points in meshes should be moved to a
common routine to be used by swapmesh, recordpoints and maybe other
future tools.