Fix for channel testcase: Temperature was not set correctly for the given mach number.
Fixes a bug that was introduced with the option to specify a mach number
for the channel test case. While the pressure was set correctly in the
reference state, the corresponding temperature was not calculated. This
means that for isothermal walls, the wrong temperature was used in the
boundary fluxes, chrashing the simulation.