Compatibility of EddyVisc and Posti
Fix bugs appearing when trying to run posti with eddy viscosity enabled.
First, the DefineParameters was not called in ReadState. This is fixed
by moving the DefineParameters of the EddyVisc module to the respective
call in the Equation module, where it should belong anyway. Second, a
segmentation fault occured when calling the FinalizeEquation routine
because of the embedded call to FinalizeEddyViscosity, a function
pointer that may not be set yet. To circumvent that, add an ASSOCIATE
check.