Allow posti_visu to visualize files with different node types
In case the equation systems match, directly call the normal restart routines. The other case is rather annoying as we have no information during compile time on the number of variables in the state file. Provide a copy of the interpolation routine embedded in Restart(), but adapt it to handle an arbitrary number of variables. Tested with State, TimeAvg, and BaseFlow files and seems to be working.