Fortidude: Fix T041: assumed-size
Assumed size dummy arguments declared with a star  as the size should be avoided. There are several downsides to assumed size, the main one being that the compiler is not able to determine the array bounds, so it is not possible to check for array overruns or to use the array in whole-array expressions.